+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/spring/SpringBooter.java
@@ -19,6 +19,7 @@ import com.ericsson.bss.cassandra.ecchronos.rest.OnDemandRepairManagementRESTImp
 import com.ericsson.bss.cassandra.ecchronos.rest.RepairManagementRESTImpl;
 import com.ericsson.bss.cassandra.ecchronos.rest.ScheduleRepairManagementRESTImpl;
 import com.ericsson.bss.cassandra.ecchronos.rest.RejectConfigREST;
+import com.ericsson.bss.cassandra.ecchronos.rest.StateRepairManagementRESTImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
@@ -28,7 +29,8 @@ import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
 @Import(value = {RepairManagementRESTImpl.class, ScheduleRepairManagementRESTImpl.class,
-        OnDemandRepairManagementRESTImpl.class, MetricsREST.class, RejectConfigREST.class})
+        OnDemandRepairManagementRESTImpl.class, MetricsREST.class, RejectConfigREST.class,
+        StateRepairManagementRESTImpl.class})
 public class SpringBooter extends SpringBootServletInitializer
 {
     private static final Logger LOG = LoggerFactory.getLogger(SpringBooter.class);