+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Version 6.1.3
 
+* Rework the status commands of the ecctool for more details - Issue #946
+
 ## Version 6.1.2
 
 ## Merged from Version 6.0